Fix bug caused by passing an array as a string in Resource controller
In order to create this fix, the ResourceController has been changed to extend the ServerController, so it has access to the ServerConfig.
The ServerConfig has a getAllowedOrigins method added that returns all allowed origins it is aware of.
In order to facilitate this, some cleanup was also needed for the ServerController.